+++ backend/src/metrics/responsive_maintainer.ts
@@ -1,5 +1,5 @@
-import { graphqlRequest } from "./graphql_request";
-import logger from "./logger";
+import { graphqlRequest } from "../graphql_request";
+import logger from "../logger";
 
 export async function getResponsive(owner: string, repoName: string): Promise<number[]> {
     // Record the start time
