--- backend/src/sum.ts
@@ -1,3 +1,0 @@
-export function sum(a: number, b: number): number {
-    return a + b;
-}
